We loveee lighting and player ctrls totally
I don't see the light at the end of the tunnel, there is only darkness, the scaryness, and the fact that Unity is so annoying and area lights and baking takes forever. It did not bake, it burned the kitchen down and the cake did not survive. There are not even crumbs it's all ASHES.

(Hi Lewis I'm not insane I swear)